requirements.txt
fastapi
uvicorn

pyyaml
numpy
pandas
openpyxl
pylint
ruff
black
bandit
trufflehog3
asyncpg
pytest
coverage
mypy
httpx
pytest-asyncio